pyproject.toml

[project]
name = "openmfda_flow"
version = "0.2.1"
dependencies = [
    "solidpython>=1.1.3",
    "numpy",
    "numpy-stl>=3.1.1",
    "pandas>=1.5.2",
    "PyQt5>=5.15.10",
    "Pillow>=10.2.0",
    "matplotlib>=3.9.2",
    "regex>=2024.9.11",
    "networkx>=3.2.1",
    "lark>=1.2.2"
]
requires-python = ">=3.9"
readme = "README.md"

[tool.setuptools]
package-dir= {"" = "src"}
packages = ["openmfda_flow"]

[project.urls]
Homepage="https://github.com/utah-MFDA/openmfda_flow"
